client and status api and integration test

+++ tests/client/const.go
@@ -1,6 +1,8 @@
+// +build client_integration_tests
+
 package client
 
 const (
 	NETWORK_ID = "sebak-test-network"
-	fee        = 10000
+	fee        = uint64(10000)
 )
